requirements.txt
fastapi
openai
uvicorn
langchain
faiss-cpu  # If replacing Azure Cognitive Search
